requirements.txt
openai>=1.12.0
pydantic>=2.6.0
python-dotenv>=1.0.0
requests>=2.31.0
rich>=13.7.0
typer>=0.9.0